refactor: 💡 tailwind and use server
