+++ components.json
@@ -12,6 +12,6 @@
 	},
 	"aliases": {
 		"components": "@/components",
-		"utils": "@/lib/utils"
+		"utils": "@/utils/tailwindcss"
 	}
 }
