+++ src/components/ui/select.tsx
@@ -9,7 +9,7 @@ import {
 import * as SelectPrimitive from "@radix-ui/react-select";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/tailwindcss";
 
 const Select = SelectPrimitive.Root;
