+++ src/features/auth/actions/sign-out.tsx
@@ -1,4 +1,5 @@
 "use server";
+import "server-only";
 import { sendLineNotifyMessage } from "@/apis/line-notify/send-message";
 import { ERROR_MESSAGES, SUCCESS_MESSAGES } from "@/constants";
 import { signOut as NextAuthSignOut } from "@/features/auth/lib/auth";
