+++ src/features/dump/actions/type.ts
@@ -1,3 +1,5 @@
+"use server";
+import "server-only";
 import type { ServerAction } from "@/types";
 
 export type ActionState<T> =
