+++ src/features/submit/actions/change-news-status.ts
@@ -1,12 +1,12 @@
 "use server";
 import "server-only";
-import { LineNotifyError } from "@/apis/line-notify/error";
-import { formatChangeStatusMessage } from "@/apis/line-notify/format-for-line";
 import { sendLineNotifyMessage } from "@/apis/line-notify/send-message";
 import { revertNewsStatus, updateNewsStatus } from "@/apis/prisma/fetch-news";
 import { ERROR_MESSAGES } from "@/constants";
+import { LineNotifyError } from "@/error";
 import { auth } from "@/features/auth/lib/auth";
 import type { ServerAction } from "@/types";
+import { formatChangeStatusMessage } from "@/utils/format-for-line";
 import { Prisma } from "@prisma/client";
 
 type Change = "UPDATE" | "REVERT";
